requirements.txt
Django>=5.1,<5.3
djangorestframework>=3.15,<3.16
Pillow>=10.2
qrcode>=7.4
coverage>=7.5
reportlab>=4.0
dj-database-url
psycopg[binary]
